Initial work on split implementation similar to upstream st
crude first steps to move the wayland stack to a state where it becomes
easy to track upstream st. Most wayland related functions are separated
out and the program compiles and runs. Further work is needed to flesh
out bugs, clean up the messy code and eventually merge with upstream.